+++ uui-timeline/src/draw/scale.ts
@@ -14,6 +14,8 @@ import {
     CanvasScaleRange,
     CanvasDrawBottomGridLine,
     CanvasDrawWeekendHoursCell,
+    CanvasDrawBottomMonthProps,
+    CanvasDrawTopMonthProps,
 } from './types';
 
 const defaultFonts = {
@@ -36,6 +38,7 @@ const defaultColors = {
 
 const defaultWidth = {
     cellBorderWidth: 1,
+    todayLineHeight: 4,
 };
 
 const moveAmount = 0.7;
@@ -45,6 +48,8 @@ const isCurrentPeriod = (leftDate: Date, rightDate: Date) => new Date() >= leftD
 
 const getCanvasVerticalCenter = (canvasHeight: number) => canvasHeight / 2 - 1;
 const getBottomCellY = (canvasHeight: number) => getCanvasVerticalCenter(canvasHeight);
+const getTopMonth = (month: number) => months[month]?.toUpperCase() ?? '';
+const getBottomMonth = (month: number) => months[month] ?? '';
 
 const drawScaleBottomBorder = ({
     context,
@@ -375,10 +380,16 @@ const drawHours = ({
         });
 };
 
-const drawToday = ({ context, scaleBar, todayLineColor = defaultColors.todayLineColor }: CanvasDrawHeaderTodayProps) => {
+const drawToday = ({
+    context,
+    scaleBar,
+    todayLineColor = defaultColors.todayLineColor,
+    canvasHeight,
+    todayLineHeight = defaultWidth.todayLineHeight,
+}: CanvasDrawHeaderTodayProps) => {
     if (isCurrentPeriod(scaleBar.leftDate, scaleBar.rightDate)) {
         context.fillStyle = todayLineColor;
-        context.fillRect(scaleBar.left + 1, 56, scaleBar.right - scaleBar.left, 4);
+        context.fillRect(scaleBar.left + 1, canvasHeight - todayLineHeight, scaleBar.right - scaleBar.left, todayLineHeight);
     }
 };
 
@@ -389,7 +400,9 @@ const drawTopDays = ({
     topDayTextColor = defaultColors.topDayTextColor,
     weekendTextColor = defaultColors.weekendTextColor,
     todayLineColor = defaultColors.todayLineColor,
+    todayLineHeight = defaultWidth.todayLineHeight,
     drawToday: customDrawToday,
+    getTopMonth: customGetMonth = getTopMonth,
     canvasHeight,
     cellBorderColor = defaultColors.cellBorderColor,
     cellBorderWidth = defaultWidth.cellBorderWidth,
@@ -398,7 +411,7 @@ const drawTopDays = ({
     ...restProps
 }: CanvasDrawTopDaysProps) => {
     timelineTransform.getVisibleDays().forEach((w) => {
-        const header = months[w.leftDate.getMonth()] + ' ' + w.leftDate.getDate().toString() + ', ' + w.leftDate.getFullYear();
+        const header = customGetMonth(w.leftDate.getMonth()) + ' ' + w.leftDate.getDate().toString() + ', ' + w.leftDate.getFullYear();
         const isHoliday = timelineTransform.isWeekend(w.leftDate) || timelineTransform.isHoliday(w.leftDate);
         const color = isHoliday ? weekendCellBackgroundColor : cellBackgroundColor;
         drawCellBackground({ context, scaleBar: w, canvasHeight, color });
@@ -408,7 +421,7 @@ const drawTopDays = ({
         drawPeriodText({
             context,
             timelineTransform,
-            text: header.toUpperCase(),
+            text: header,
             x: w.left,
             width: w.right - w.left,
             line: getTopLine(visibility),
@@ -416,7 +429,7 @@ const drawTopDays = ({
             textColor,
             ...restProps,
         });
-        (customDrawToday ?? drawToday)({ context, scaleBar: w, todayLineColor });
+        (customDrawToday ?? drawToday)({ context, scaleBar: w, todayLineColor, todayLineHeight, canvasHeight });
         drawBorderForTopCell({ context, canvasHeight, scaleBar: w, width: cellBorderWidth, color: cellBorderColor });
     });
 };
@@ -428,6 +441,7 @@ const drawDays = ({
     periodTextColor = defaultColors.periodTextColor,
     weekendTextColor = defaultColors.weekendTextColor,
     todayLineColor = defaultColors.todayLineColor,
+    todayLineHeight = defaultWidth.todayLineHeight,
     drawToday: customDrawToday,
     canvasHeight,
     cellBorderColor = defaultColors.cellBorderColor,
@@ -442,7 +456,7 @@ const drawDays = ({
         const color = isHoliday ? weekendCellBackgroundColor : cellBackgroundColor;
         drawCellBackground({ context, scaleBar: w, canvasHeight, y: getCanvasVerticalCenter(canvasHeight), color });
 
-        (customDrawToday ?? drawToday)({ context, scaleBar: w, todayLineColor });
+        (customDrawToday ?? drawToday)({ context, scaleBar: w, todayLineColor, todayLineHeight, canvasHeight });
         drawBottomGridLine({ context, canvasHeight, scaleBar: w, width: cellBorderWidth, color: cellBorderColor });
 
         const text = w.leftDate.getDate().toString();
@@ -472,8 +486,9 @@ const drawTopMonths = ({
     cellBorderWidth = defaultWidth.cellBorderWidth,
     cellBackgroundColor = defaultColors.cellBackgroundColor,
     evenPeriodCellBackgroundColor = defaultColors.evenPeriodCellBackgroundColor,
+    getTopMonth: customGetMonth = getTopMonth,
     ...restProps
-}: CanvasDrawPeriodPartProps) => {
+}: CanvasDrawTopMonthProps) => {
     timelineTransform.getVisibleMonths().forEach((w) => {
         const color = w.leftDate.getMonth() % 2 === 0
             ? cellBackgroundColor
@@ -482,12 +497,12 @@ const drawTopMonths = ({
         drawCellBackground({ context, scaleBar: w, canvasHeight, color });
         drawBorderForTopCell({ context, canvasHeight, scaleBar: w, width: cellBorderWidth, color: cellBorderColor });
 
-        const header = months[w.leftDate.getMonth()] + ' ' + w.leftDate.getFullYear();
+        const header = customGetMonth(w.leftDate.getMonth()) + ' ' + w.leftDate.getFullYear();
         const isCurPeriod = isCurrentPeriod(w.leftDate, w.rightDate);
         drawPeriodText({
             context,
             timelineTransform,
-            text: header.toUpperCase(),
+            text: header,
             x: w.left,
             width: w.right - w.left,
             line: getTopLine(visibility),
@@ -504,6 +519,7 @@ const drawWeeks = ({
     visibility,
     periodTextColor = defaultColors.periodTextColor,
     todayLineColor = defaultColors.todayLineColor,
+    todayLineHeight = defaultWidth.todayLineHeight,
     drawToday: customDrawToday,
     canvasHeight,
     cellBorderColor = defaultColors.cellBorderColor,
@@ -517,7 +533,7 @@ const drawWeeks = ({
         const isCurPeriod = isCurrentPeriod(w.leftDate, w.rightDate);
         drawCellBackground({ context, scaleBar: w, canvasHeight, y: getCanvasVerticalCenter(canvasHeight), color: cellBackgroundColor });
 
-        (customDrawToday ?? drawToday)({ context, scaleBar: w, todayLineColor });
+        (customDrawToday ?? drawToday)({ context, scaleBar: w, todayLineColor, todayLineHeight, canvasHeight });
         drawBottomGridLine({ context, canvasHeight, scaleBar: w, width: cellBorderWidth, color: cellBorderColor });
         drawPeriodText({
             context,
@@ -540,15 +556,16 @@ const drawBottomMonths = ({
     canvasHeight,
     periodTextColor = defaultColors.periodTextColor,
     todayLineColor = defaultColors.todayLineColor,
+    todayLineHeight = defaultWidth.todayLineHeight,
     drawToday: customDrawToday,
     cellBorderColor = defaultColors.cellBorderColor,
     cellBorderWidth = defaultWidth.cellBorderWidth,
     cellBackgroundColor = defaultColors.cellBackgroundColor,
-
+    getBottomMonth: customGetMonth = getBottomMonth,
     ...restProps
-}: CanvasDrawPeriodWithTodayProps) => {
+}: CanvasDrawBottomMonthProps) => {
     timelineTransform.getVisibleMonths().forEach((w) => {
-        const text = months[w.leftDate.getMonth()].toString();
+        const text = customGetMonth(w.leftDate.getMonth()).toString();
         const isCurPeriod = isCurrentPeriod(w.leftDate, w.rightDate);
         drawCellBackground({ context, scaleBar: w, canvasHeight, y: getCanvasVerticalCenter(canvasHeight), color: cellBackgroundColor });
 
@@ -563,7 +580,7 @@ const drawBottomMonths = ({
             textColor: periodTextColor,
             ...restProps,
         });
-        (customDrawToday ?? drawToday)({ context, scaleBar: w, todayLineColor });
+        (customDrawToday ?? drawToday)({ context, scaleBar: w, todayLineColor, todayLineHeight, canvasHeight });
         drawBottomGridLine({ context, canvasHeight, scaleBar: w, width: cellBorderWidth, color: cellBorderColor });
     });
 };
@@ -574,6 +591,7 @@ const drawYears = ({
     visibility,
     periodTextColor = defaultColors.periodTextColor,
     todayLineColor = defaultColors.todayLineColor,
+    todayLineHeight = defaultWidth.todayLineHeight,
     drawToday: customDrawToday,
     canvasHeight,
     cellBorderColor = defaultColors.cellBorderColor,
@@ -593,7 +611,7 @@ const drawYears = ({
             const y = canvasHeight;
             timelinePrimitives.drawHorizontalLine({ context, x1: w.left, x2: w.right + 1, y: y - 1, color: cellBorderColor, width: cellBorderWidth });
             drawCellBackground({ context, scaleBar: w, canvasHeight, height: canvasHeight, color: cellBackgroundColor });
-            (customDrawToday ?? drawToday)({ context, scaleBar: w, todayLineColor });
+            (customDrawToday ?? drawToday)({ context, scaleBar: w, todayLineColor, todayLineHeight, canvasHeight });
             timelinePrimitives.drawVerticalLine({ context, x: w.left + 0.5, y2: y - 1, color: cellBorderColor, width: cellBorderWidth });
         } else {
             const color = w.leftDate.getFullYear() % 2 === 0
@@ -657,7 +675,8 @@ export const timelineScale = {
     getBottomMonthsScaleRange,
     getYearsScaleRange,
     drawHoursCells,
-
+    getTopMonth,
+    getBottomMonth,
     defaultFonts,
     defaultColors,
     defaultWidth,